# apps/app/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SkyPlanner API
  description: |
    REST API for SkyPlanner kundeadministrasjon og ruteplanlegging.

    ## Autentisering

    API-et stotter to autentiseringsmetoder:

    ### API-nokkel
    For tredjeparts-integrasjoner. Opprett API-nokkel i admin-panelet.
    ```
    X-API-Key: sk_live_xxx
    ```

    ### JWT Bearer Token
    For innloggede brukere via web-applikasjonen.
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - Standard: 100 requests per 15 minutter
    - API-nokler kan ha egne kvote-innstillinger

    ## Feilhandtering
    Alle feil returnerer et standardisert format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Beskrivende feilmelding"
      },
      "requestId": "abc123"
    }
    ```

  version: 1.0.0
  contact:
    name: Efffekt AS
    email: support@skyplanner.no
    url: https://skyplanner.no
servers:
  - url: https://app.skyplanner.no/api
    description: Produksjonsserver
  - url: http://localhost:3000/api
    description: Lokal utvikling
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api-keys:
    get:
      tags:
        - API-nokler
      summary: List API-nokler
      description: Henter alle API-nokler for organisasjonen
      operationId: listApiKeys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste over API-nokler
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApiKey'

    post:
      tags:
        - API-nokler
      summary: Opprett API-nokkel
      description: |
        Oppretter en ny API-nokkel.

        **Viktig:** Nokkel-verdien returneres kun ved opprettelse og må lagres trygt.
      operationId: createApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API-nokkel opprettet
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          apiKey:
                            $ref: '#/components/schemas/ApiKey'
                          key:
                            type: string
                            description: Den fulle API-nokkelen (vises kun en gang)
                            example: sk_live_abc123...
  /api-keys/{id}/revoke:
    post:
      tags:
        - API-nokler
      summary: Deaktiver API-nokkel
      operationId: revokeApiKey
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Årsak til deaktivering
      responses:
        '200':
          description: API-nokkel deaktivert

  # ============ Health ============
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-nokkel opprettet i admin-panelet

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-token fra innlogging
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Unik ID
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          example: req_abc123
...
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        key_prefix:
          type: string
          example: sk_live_abc
          description: Prefix for identifikasjon
        name:
          type: string
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
            enum:
              - customers:read
              - customers:write
              - routes:read
              - routes:write
              - appointments:read
              - appointments:write
              - webhooks:manage
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateApiKeyRequest:
      type: object
      required:
        - name
        - scopes
      properties:
        name:
          type: string
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
          minItems: 1
        expires_at:
          type: string
          format: date-time
          description: Utlopssdato (valgfri)
        monthly_quota:
          type: integer
          description: Maks antall requests per måned
        rate_limit_requests:
          type: integer
          description: Maks requests per vindu
        rate_limit_window_seconds:
          type: integer
          description: Tidsvindu for rate limiting
...